assets/requirements.txt
opencv-python
numpy
pygame
twilio
tensorflow
python-dotenv